summarization/requirements.txt
streamlit
openai
python-dotenv
requests
pandas
ipython
beautifulsoup4